refactor(core): fold reply into message
